casting an int to a float seems silly, if its going to be auto promoted to a double anyway.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@9129 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/BackgroundRenderer.cxx
@@ -866,7 +866,7 @@ void BackgroundRenderer::drawGroundReceivers(SceneRenderer& renderer)
 
   if (!init) {
     init = true;
-    const float receiverSliceAngle = (float)(2.0 * M_PI / float(receiverSlices));
+    const float receiverSliceAngle = (float)(2.0 * M_PI / double(receiverSlices));
     for (int i = 0; i <= receiverSlices; i++) {
       angle[i][0] = cosf((float)i * receiverSliceAngle);
       angle[i][1] = sinf((float)i * receiverSliceAngle);
